refactor: Use toolbar container + section instead of css
