--- application/ui/src/features/dataset/media-preview/media-preview.module.scss
@@ -1,18 +1,0 @@
-.toolbarGrid {
-    gap: var(--spectrum-global-dimension-size-50);
-    background-color: var(--spectrum-global-color-gray-200);
-    border-radius: var(--spectrum-global-dimension-size-100);
-    padding: var(--spectrum-global-dimension-size-50);
-}
-
-.toolbarSection {
-    background-color: var(--spectrum-global-color-gray-50);
-    padding: var(--spectrum-global-dimension-size-100);
-    border-radius: var(--spectrum-global-dimension-size-50);
-    gap: var(--spectrum-global-dimension-size-50);
-    align-items: center;
-}
-
-.filename {
-    font-size: var(--spectrum-global-dimension-size-150);
-}
