+++ application/ui/src/features/dataset/media-preview/secondary-toolbar/secondary-toolbar.component.tsx
@@ -11,10 +11,10 @@ import { useAnnotationActions } from '../../../../shared/annotator/annotation-ac
 import { useAnnotator } from '../../../../shared/annotator/annotator-provider.component';
 import { useSelectedAnnotations } from '../../../../shared/annotator/select-annotation-provider.component';
 import { DeleteMediaItem } from '../../gallery/delete-media-item/delete-media-item.component';
+import { Toolbar } from '../toolbar-container/toolbar-container.component';
 import { AnnotatorModes } from './annotator-modes/annotator-modes-toggle.component';
 import type { AnnotatorMode } from './annotator-modes/mode';
 import { LabelPicker } from './label-picker.component';
-import { ToolbarContainer } from './toolbar-container/toolbar-container.component';
 import { useSecondaryToolbarState } from './use-secondary-toolbar-state.hook';
 
 import styles from './secondary-toolbar.module.scss';
@@ -97,47 +97,53 @@ export const SecondaryToolbar = ({
             justifyContent={'space-between'}
             UNSAFE_style={{ paddingTop: dimensionValue('size-125') }}
         >
-            <ToolbarContainer>
-                <AnnotatorModes mode={mode} onModeChange={onModeChange} />
-            </ToolbarContainer>
-            <ToolbarContainer isHidden={isHidden}>
-                <LabelPicker selectedLabel={selectedLabel} labels={projectLabels} onSelect={handleSelect} />
-            </ToolbarContainer>
-            <ToolbarContainer>
-                <ButtonGroup>
-                    <DeleteMediaItem
-                        itemsIds={[String(mediaItem.id)]}
-                        onDeleted={([deletedItem]: string[]) => handleDeleteItem([deletedItem], items.length - 1)}
-                    />
-                    <Button
-                        variant='accent'
-                        onPress={handleSubmit}
-                        isPending={isSaving}
-                        marginStart={'size-200'}
-                        isDisabled={!hasAnnotations || isSaving}
-                    >
-                        {mode === 'annotation' ? (
-                            'Submit'
-                        ) : (
-                            <>
-                                <Checkmark />
-                                <Text>Confirm prediction</Text>
-                            </>
-                        )}
-                    </Button>
+            <Toolbar.Container>
+                <Toolbar.Section>
+                    <AnnotatorModes mode={mode} onModeChange={onModeChange} />
+                </Toolbar.Section>
+            </Toolbar.Container>
+            <Toolbar.Container isHidden={isHidden}>
+                <Toolbar.Section>
+                    <LabelPicker selectedLabel={selectedLabel} labels={projectLabels} onSelect={handleSelect} />
+                </Toolbar.Section>
+            </Toolbar.Container>
+            <Toolbar.Container>
+                <Toolbar.Section>
+                    <ButtonGroup>
+                        <DeleteMediaItem
+                            itemsIds={[String(mediaItem.id)]}
+                            onDeleted={([deletedItem]: string[]) => handleDeleteItem([deletedItem], items.length - 1)}
+                        />
+                        <Button
+                            variant='accent'
+                            onPress={handleSubmit}
+                            isPending={isSaving}
+                            marginStart={'size-200'}
+                            isDisabled={!hasAnnotations || isSaving}
+                        >
+                            {mode === 'annotation' ? (
+                                'Submit'
+                            ) : (
+                                <>
+                                    <Checkmark />
+                                    <Text>Confirm prediction</Text>
+                                </>
+                            )}
+                        </Button>
 
-                    <ActionButton
-                        isQuiet
-                        onPress={onClose}
-                        isDisabled={isSaving}
-                        marginStart={'size-100'}
-                        UNSAFE_className={styles.closeButton}
-                    >
-                        <CloseSemiBold width={14} height={14} />
-                        <Text>Close</Text>
-                    </ActionButton>
-                </ButtonGroup>
-            </ToolbarContainer>
+                        <ActionButton
+                            isQuiet
+                            onPress={onClose}
+                            isDisabled={isSaving}
+                            marginStart={'size-100'}
+                            UNSAFE_className={styles.closeButton}
+                        >
+                            <CloseSemiBold width={14} height={14} />
+                            <Text>Close</Text>
+                        </ActionButton>
+                    </ButtonGroup>
+                </Toolbar.Section>
+            </Toolbar.Container>
         </Flex>
     );
 };
